requirements.txt
clearml==1.18.0
albumentations
pillow==11.0.0
pandas==2.2.3
plotly==5.23.0
kaleido==0.2.1
nbformat==5.10.4
lightning==2.5.1
pytorch-lightning==2.5.1
tensorboard==2.19.0
opencv-python==4.11.0.86
timm==1.0.15
lmdb==1.6.2
# torch==2.7.0+cu126
# torchvision==0.22.0+cu126
